ENG-11743 Addressed few review comments.
