+++ NodeJS/src/AgentSkill.ts
@@ -1,19 +1,13 @@
 import { createSession } from "./session"
 import { createAgentClient } from "./AgentClient"
 import * as Constants from "./Constants";
+import isValidParameter from "./Utils";
+
 const args = require('minimist')(process.argv.slice(2));
 
 let endpoint = args[Constants.ENDPOINT_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
 let adminUsername = args[Constants.ADMIN_USERNAME_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
 let adminPassword = args[Constants.ADMIN_PASSWORD_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
-
-function isValidParameter(key: string, parameter: undefined): boolean {
-    if (!parameter) {
-        console.log(key + ' was undefined');
-        return false;
-    }
-    return true;
-}
 
 let isEndpointValid = isValidParameter(Constants.ENDPOINT_KEY, endpoint);
 let isAdminUsernameValid = isValidParameter(Constants.ADMIN_USERNAME_KEY, adminUsername);
