+++ NodeJS/src/SampleAgentClient.ts
@@ -1,8 +1,8 @@
 import {Err, Ok, Result} from "ts.data.json/dist/result";
-import { JsonDecoder } from 'ts.data.json';
 import * as Constants from "./Constants";
 import { createSession } from "./session";
 import SkillPriority, { createAgentClient} from "./AgentClient";
+import isValidParameter,{ isValidSkillsWithPriorities} from "./Utils";
 
 const args = require('minimist')(process.argv.slice(2));
 
@@ -12,32 +12,6 @@ let adminPassword = args[Constants.ADMIN_PASSWORD_KEY].replace(Constants.REPLACE
 let agentUsername = args[Constants.AGENT_USERNAME_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
 let agentPassword = args[Constants.AGENT_PASSWORD_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
 let agentSkill = args[Constants.AGENT_SKILL_KEY].replace(Constants.REPLACE_REGEX, Constants.EMPTY_STRING);
-
-function isValidParameter(key: string, parameter: undefined): boolean {
-    if (!parameter) {
-        console.log(key + ' was undefined');
-        return false;
-    }
-    return true;
-}
-
-const skillDecoder = JsonDecoder.object<SkillPriority>(
-    {
-        skillNumber: JsonDecoder.number,
-        skillPriority: JsonDecoder.number
-    },
-    'Skill Priority Combination'
-);
-
-function isValidSkillsWithPriorities(key: string, skillPriorities: string): boolean {
-    let obj: [SkillPriority]  = JSON.parse(skillPriorities);
-    for(let skill of obj){
-        if(skillDecoder.decode(skill) instanceof Err){
-            return false;
-        }
-    }
-    return true;
-}
 
 let isSkillValid = isValidSkillsWithPriorities(Constants.AGENT_SKILL_KEY,agentSkill)
 let isEndpointValid = isValidParameter(Constants.ENDPOINT_KEY, endpoint);
